chore: move 2 migrations to sdl pattern
