+++ packages/server/graphql/rootMutation.ts
@@ -9,7 +9,6 @@ import addPokerTemplateDimension from './mutations/addPokerTemplateDimension'
 import addPokerTemplateScale from './mutations/addPokerTemplateScale'
 import addPokerTemplateScaleValue from './mutations/addPokerTemplateScaleValue'
 import addReflectTemplatePrompt from './mutations/addReflectTemplatePrompt'
-import addSlackAuth from './mutations/addSlackAuth'
 import addTeam from './mutations/addTeam'
 import archiveOrganization from './mutations/archiveOrganization'
 import archiveTeam from './mutations/archiveTeam'
@@ -83,7 +82,6 @@ import resetPassword from './mutations/resetPassword'
 import resetRetroMeetingToGroupStage from './mutations/resetRetroMeetingToGroupStage'
 import selectTemplate from './mutations/selectTemplate'
 import setAppLocation from './mutations/setAppLocation'
-import setDefaultSlackChannel from './mutations/setDefaultSlackChannel'
 import setNotificationStatus from './mutations/setNotificationStatus'
 import setPhaseFocus from './mutations/setPhaseFocus'
 import setPokerSpectate from './mutations/setPokerSpectate'
@@ -122,7 +120,6 @@ export default new GraphQLObjectType<any, GQLContext>({
       addPokerTemplateScale,
       addPokerTemplateScaleValue,
       addReflectTemplatePrompt,
-      addSlackAuth,
       addGitHubAuth,
       addOrg,
       addTeam,
@@ -189,7 +186,6 @@ export default new GraphQLObjectType<any, GQLContext>({
       resetRetroMeetingToGroupStage,
       selectTemplate,
       setAppLocation,
-      setDefaultSlackChannel,
       setPhaseFocus,
       setStageTimer,
       setSlackNotification,
